+++ src/components/CategoryWidget.jsx
