+++ src/components/Sorters.jsx
@@ -1,5 +1,3 @@
-import { useState } from "react";
-
 const Sorters = ({
   setSelectedOrder,
   setSelectedSortBy,
